+++ src/main/java/com/twilio/rest/authy/v1/service/entity/FactorCreator.java
@@ -28,6 +28,7 @@ public class FactorCreator extends Creator<Factor> {
     private final String binding;
     private final String friendlyName;
     private final Factor.FactorTypes factorType;
+    private final String config;
 
     /**
      * Construct a new FactorCreator.
@@ -37,17 +38,20 @@ public class FactorCreator extends Creator<Factor> {
      * @param binding A unique binding for this Factor as a json string
      * @param friendlyName The friendly name of this Factor
      * @param factorType The Type of this Factor
+     * @param config The config for this Factor as a json string
      */
     public FactorCreator(final String pathServiceSid,
                          final String pathIdentity,
                          final String binding,
                          final String friendlyName,
-                         final Factor.FactorTypes factorType) {
+                         final Factor.FactorTypes factorType,
+                         final String config) {
         this.pathServiceSid = pathServiceSid;
         this.pathIdentity = pathIdentity;
         this.binding = binding;
         this.friendlyName = friendlyName;
         this.factorType = factorType;
+        this.config = config;
     }
 
     /**
@@ -105,6 +109,10 @@ public class FactorCreator extends Creator<Factor> {
 
         if (factorType != null) {
             request.addPostParam("FactorType", factorType.toString());
+        }
+
+        if (config != null) {
+            request.addPostParam("Config", config);
         }
     }
 }